refactor: restructure components into subfolders with readmes
- organize each component into its own subfolder
- add style prop to avatar and badge components
- create readme.md files with api documentation in table format
- update imports and exports to match new structure
