+++ components/index.ts
@@ -5,21 +5,21 @@
  */
 
 // Layout
-export { Box, type BoxProps } from "./box"
-export { Flex, type FlexProps } from "./flex"
-export { Card, type CardProps } from "./card"
-export { Divider, type DividerProps } from "./divider"
+export { Box, type BoxProps } from "./box/box"
+export { Flex, type FlexProps } from "./flex/flex"
+export { Card, type CardProps } from "./card/card"
+export { Divider, type DividerProps } from "./divider/divider"
 
 // Typography
-export { Text, type TextProps } from "./text"
+export { Text, type TextProps } from "./text/text"
 
 // Interactive
-export { Pressable, type PressableProps } from "./pressable"
-export { Button, type ButtonProps } from "./button"
-export { Input, type InputProps } from "./input"
+export { Pressable, type PressableProps } from "./pressable/pressable"
+export { Button, type ButtonProps } from "./button/button"
+export { Input, type InputProps } from "./input/input"
 
 // Display
-export { Icon, type IconProps } from "./icon"
-export { Badge, type BadgeProps } from "./badge"
-export { Avatar, type AvatarProps } from "./avatar"
-export { Spinner, type SpinnerProps } from "./spinner"
+export { Icon, type IconProps } from "./icon/icon"
+export { Badge, type BadgeProps } from "./badge/badge"
+export { Avatar, type AvatarProps } from "./avatar/avatar"
+export { Spinner, type SpinnerProps } from "./spinner/spinner"
