+++ components/spinner/readme.md
@@ -1,0 +1,18 @@
+# Spinner
+
+Loading indicator with theme-aware colors and sizes.
+
+## api reference
+
+| prop   | type                                                                      | default | description              |
+|--------|---------------------------------------------------------------------------|---------|--------------------------|
+| size   | `sm` \| `md` \| `lg`                                                      | `md`    | spinner size             |
+| color  | `brand` \| `foreground` \| `inverse` \| `success` \| `error` \| `warning` \| `info` | `brand` | spinner color            |
+
+## usage
+
+```tsx
+import { Spinner } from "@/components"
+
+<Spinner size="lg" color="brand" />
+```
